Organize deployment test files
